First cut
This is nowhere near complete, but it *is* working. What I have right
now is the ability to respond to messages from the server on an event
loop (of a sort) and, also, the ability to send arbitrary messages
initiated by the client. I was never able to achieve that second
capability using the IRC library I had been fiddling with before
(hiirc).

I suppose I could have tried to find this capability in another irc
library, but, honestly, when I was looking at the other options I felt
as if they were pretty lacking in abstractions. Like, hooray, you're
giving me the ability to do everything for myself? What exactly is your
library for, again?

Anyway, I have a lot left to implement myself, but the shim is able to
connect to a server !!!AND!!! if you enter the commands by hand it's
able to navigate around. It's the world's worst IRC client!